ollama genAI app/requirements.txt
langchain
python-dotenv
beautifulsoup4
langchain_community
faiss-cpu
streamlit